+++ djservices/__init__.py
@@ -3,7 +3,7 @@ from .base import (
 )
 
 
-__version__ = "2.0.0"
+__version__ = "2.0.1-alpha"
 
 __all__ = [
     'BaseService', 'BaseCRUDService', 'CommonCRUDService', 'UserCRUDService'